requirements.txt
wechatpy
Django==1.8.17
django-contrib-comments==1.8.0
gevent==1.2.2
leancloud-sdk==1.12.0
